+++ packages/vue/src/Link.ts
@@ -33,24 +33,24 @@ const Link: ComponentOptions<LinkComponent> = {
 
   computed: {
     location: function() {
-      const pathname = this.$curi.config.addons.pathname(this.to, this.params);
+      const pathname = this.$curi.router.addons.pathname(this.to, this.params);
       return { pathname, ...this.details };
     },
     href: function() {
-      return this.$curi.config.history.toHref(this.location);
+      return this.$curi.router.history.toHref(this.location);
     },
     isActive: function() {
-      if (!this.$curi.config.addons.active) {
+      if (!this.$curi.router.addons.active) {
         warning(
           !this.active,
           'You are attempting to use the "active" prop, but have not included the "active" ' +
-            'addon (@curi/addon-active) in your Curi configuration object.'
+            'addon (@curi/addon-active) in your Curi router.'
         );
         return false;
       }
 
       return (
-        this.$curi.config.addons.active(
+        this.$curi.router.addons.active(
           this.to,
           this.$curi.response,
           this.params,
@@ -71,7 +71,7 @@ const Link: ComponentOptions<LinkComponent> = {
 
       if (canNavigate(event)) {
         event.preventDefault();
-        this.$curi.config.history.navigate(this.location);
+        this.$curi.router.history.navigate(this.location);
       }
     }
   },