+++ packages/vue/src/interface.ts
@@ -1,9 +1,9 @@
 import Vue from 'vue';
-import { CuriConfig, Response } from '@curi/core';
+import { CuriRouter, Response } from '@curi/core';
 import { Action } from '@hickory/root';
 
 export interface ReactiveCuriProps {
-  config: CuriConfig;
+  router: CuriRouter;
   response: Response;
   action: Action;
 }