+++ packages/vue/tests/plugin.spec.ts
@@ -1,6 +1,6 @@
 import 'jest';
 import { createLocalVue, shallow, mount } from 'vue-test-utils';
-import createConfig from '@curi/core';
+import curi from '@curi/core';
 import InMemory from '@hickory/in-memory';
 import CuriPlugin from '../src/plugin';
 import reactiveCuri from '../src/reactive';
@@ -8,7 +8,7 @@ import reactiveCuri from '../src/reactive';
 describe('CuriPlugin', () => {
   const history = InMemory();
   const routes = [];
-  const config = createConfig(history, routes);
+  const router = curi(history, routes);
 
   describe('$curi', () => {
     it('Adds a mixin that sets $curi property for all components', done => {
@@ -18,16 +18,16 @@ describe('CuriPlugin', () => {
           return h('div');
         }
       };
-      config.respond(
+      router.respond(
         (response, action) => {
-          const curi = reactiveCuri(config);
+          const curi = reactiveCuri(router);
           Vue.use(CuriPlugin, { curi });
 
           const wrapper = shallow(FakeComponent, {
             localVue: Vue
           });
 
-          expect(wrapper.vm.$curi.config).toBe(config);
+          expect(wrapper.vm.$curi.router).toBe(router);
           expect(wrapper.vm.$curi.response).toBe(null);
           expect(wrapper.vm.$curi.action).toBe(null);
 
@@ -54,12 +54,12 @@ describe('CuriPlugin', () => {
         };
       }
 
-      let history, config;
+      let history, router;
       const routes = [];
 
       beforeEach(() => {
         history = InMemory();
-        config = createConfig(history, routes);
+        router = curi(history, routes);
       });
 
       it('re-renders when updating curi object', done => {
@@ -67,10 +67,10 @@ describe('CuriPlugin', () => {
         const FakeComponent = makeFake(done);
 
         let wrapper;
-        const curi = reactiveCuri(config);
+        const curi = reactiveCuri(router);
         Vue.use(CuriPlugin, { curi });
 
-        config.respond((response, action) => {
+        router.respond((response, action) => {
           curi.response = response;
           curi.action = action;
 
@@ -78,7 +78,7 @@ describe('CuriPlugin', () => {
             wrapper = shallow(FakeComponent, {
               localVue: Vue
             });
-            config.history.push('/another-one');
+            router.history.push('/another-one');
           }
         });
       });
@@ -88,10 +88,10 @@ describe('CuriPlugin', () => {
         const FakeComponent = makeFake(done);
 
         let wrapper;
-        const curi = reactiveCuri(config);
+        const curi = reactiveCuri(router);
         Vue.use(CuriPlugin, { curi });
 
-        config.respond((response, action) => {
+        router.respond((response, action) => {
           curi.response = response;
           curi.action = action;
           if (!wrapper) {
@@ -104,7 +104,7 @@ describe('CuriPlugin', () => {
                 localVue: Vue
               }
             );
-            config.history.push('/another-one');
+            router.history.push('/another-one');
           }
         });
       });
@@ -115,7 +115,7 @@ describe('CuriPlugin', () => {
     it('Registers the Link component as <curi-link>', () => {
       const Vue = createLocalVue();
       Vue.use(CuriPlugin, {
-        curi: { config, response: null, action: null }
+        curi: { router, response: null, action: null }
       });
       expect(Vue.options.components['curi-link']).toBeDefined();
     });
@@ -125,7 +125,7 @@ describe('CuriPlugin', () => {
     it('Registers the Block component as <curi-block>', () => {
       const Vue = createLocalVue();
       Vue.use(CuriPlugin, {
-        curi: { config, response: null, action: null }
+        curi: { router, response: null, action: null }
       });
       expect(Vue.options.components['curi-block']).toBeDefined();
     });