+++ packages/vue/types/install.d.ts
@@ -1,3 +1,3 @@
 import Vue from 'vue';
-import { CuriConfig } from '@curi/core';
-export default function install(_Vue: typeof Vue, config: CuriConfig): void;
+import { CuriRouter } from '@curi/core';
+export default function install(_Vue: typeof Vue, router: CuriRouter): void;