+++ packages/vue/types/reactive.d.ts
@@ -1,4 +1,4 @@
 import Vue from 'vue';
-import { CuriConfig } from '@curi/core';
+import { CuriRouter } from '@curi/core';
 import { ReactiveCuriProps } from './interface';
-export default function reactive(config: CuriConfig): ReactiveCuriProps & object & Record<never, any> & Vue;
+export default function reactive(router: CuriRouter): ReactiveCuriProps & object & Record<never, any> & Vue;